final push all modified files are push

+++ routes/review.js
@@ -6,6 +6,7 @@ const {reviewSchema}=require("../schema.js");
 const wrapAsync=require("../utils/wrapAsync.js");
 const ExpressError=require("../utils/ExpresError.js");
 const Listing=require("../models/listing.js");
+const {isLoggedIn}=require("../middleware.js");
 
 const validateReview=(req,res,next)=>{
   let{error}=reviewSchema.validate(req.body);
@@ -18,11 +19,11 @@ const validateReview=(req,res,next)=>{
   }
 }
 
-router.post("/",validateReview,async(req,res) =>{
+router.post("/",validateReview,isLoggedIn,async(req,res) =>{
   const {id}=req.params;
   const listing=await Listing.findById(id);
   const newReview=new Review(req.body.review);
-
+  newReview.author=req.user._id;
   listing.reviews.push(newReview);
 
   await newReview.save();
@@ -32,9 +33,14 @@ router.post("/",validateReview,async(req,res) =>{
   res.redirect(`/listings/${id}`);
 })
 
-router.delete("/:reviewId",async (req,res) => {
+router.delete("/:reviewId",isLoggedIn,async (req,res) => {
 const {id,reviewId}=req.params;
-  
+const review=await Review.findById(reviewId);
+  if(!review.author.equals(req.user._id))
+  {
+    req.flash("error","You are not the author");
+    return res.redirect(`/listings/${id}`);
+  }
   await Listing.findByIdAndUpdate(id,{$pull : {reviews : reviewId}});
   await Review.findByIdAndDelete(reviewId);
   req.flash("success", "Review Deleted Successfully!");
